move logic out of __init__ and into modules
