+++ .gitignore
@@ -58,3 +58,5 @@ target/
 
 # Pyenv
 .python-version
+
+.*.swp
